third_party/WebKit/LayoutTests/fast/dom/custom/created-callback-expected.txt


This is a testharness.js-based test.
PASS transfer created callback 
PASS __proto__, :unresolved and created callback timing 
PASS other callbacks for a given element are queued during the created callback and dispatched when the created callback completes 
PASS callback is called even if the element is moved to foreign document 
Harness: the test ran to completion.
